Refined code
