--- lib/view/pages/LoginScreens/SuccessAccountScreen.dart
@@ -1,74 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
-import '../providers/theme_provider.dart';
-
-class SuccessAccountScreen extends StatelessWidget {
-  const SuccessAccountScreen({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    final themeProvider = Provider.of<ThemeProvider>(context);
-    return Scaffold(
-      backgroundColor: themeProvider.backgroundColor,
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Image.asset(
-              'assets/images/SuccessIcon.png',
-              width: 64,
-              height: 64,
-            ),
-            const SizedBox(height: 10),
-             Text(
-              'Success!',
-              style: TextStyle(
-                fontSize: 34,
-                fontWeight: FontWeight.bold,
-                color: themeProvider.textColor,
-              ),
-            ),
-            const SizedBox(height: 10),
-            const Text(
-              'You have successfully signed up',
-              style: TextStyle(
-                fontSize: 17,
-                color: Colors.white38,
-              ),
-            ),
-            const SizedBox(height: 20),
-            Container(
-              width: 250,
-              decoration: BoxDecoration(
-                borderRadius: BorderRadius.circular(10),
-                gradient: const LinearGradient(
-                  colors: [Color(0xFF9CA2E8), Color(0xFF7CABEC)],
-                  begin: Alignment.centerLeft,
-                  end: Alignment.centerRight,
-                ),
-              ),
-              child: ElevatedButton(
-                onPressed: ()=> Navigator.pushReplacementNamed(context, '/home'),
-                style: ElevatedButton.styleFrom(
-                  foregroundColor: Colors.transparent,
-                  backgroundColor: Colors.transparent,
-                  elevation: 0,
-                  shape: RoundedRectangleBorder(
-                    borderRadius: BorderRadius.circular(30),
-                  ),
-                ),
-                child: Text(
-                  'Start',
-                  style: TextStyle(
-                    color: themeProvider.textColor,
-                    fontSize: 18.0,
-                  ),
-                ),
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
